requirements.txt
aiofiles==23.1.0
aiohttp==3.8.4
aiosignal==1.3.1
aiosqlite==0.19.0
anyio==3.7.1
async-timeout==4.0.2
asyncio==3.4.3
attrs==23.1.0
beautifulsoup4==4.12.2
blinker==1.5
certifi==2023.5.7
charset-normalizer==3.2.0
click==8.1.6
dataclasses-json==0.5.9
envariables==0.1
Flask==2.3.2
frozenlist==1.3.3
fsspec==2023.6.0
greenlet==2.0.2
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.17.3
httpx==0.24.1
hypercorn==0.14.4
hyperframe==6.0.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
langchain==0.0.231
langchainplus-sdk==0.0.20
llama-index==0.7.9
MarkupSafe==2.1.3
marshmallow==3.19.0
marshmallow-enum==1.5.1
multidict==6.0.4
mypy-extensions==1.0.0
nest-asyncio==1.5.6
numexpr==2.8.4
numpy==1.25.1
openai==0.27.8
openapi-schema-pydantic==1.2.4
packaging==23.1
pandas==2.0.3
priority==2.0.0
psycopg2==2.9.6
psycopg2-binary==2.9.6
pydantic==1.10.11
python-dateutil==2.8.2
python-dotenv==1.0.0
python-telegram-bot==20.4
pytz==2023.3
PyYAML==6.0
quart==0.18.4
regex==2023.6.3
requests==2.31.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
SQLAlchemy==2.0.18
tenacity==8.2.2
tiktoken==0.4.0
tqdm==4.65.0
typing-inspect==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.26.16
Werkzeug==2.3.6
wsproto==1.2.0
yarl==1.9.2
